requirements.txt
gitdb2==2.0.6
GitPython==3.0.5
smmap2==2.0.5
